refactor: rename items for standardization and improve doc
