+++ packages/core/src/application/index.ts
@@ -1,2 +1,2 @@
-export { AppContainer } from "./app-container";
+export { AppContainer } from "./application-container";
 export { AppFactory } from "./application-factory";
