+++ packages/core/src/common/color-codes.ts
