+++ packages/core/src/common/package-resolver.ts
