+++ packages/core/src/common/project-config.ts
