+++ packages/core/src/provider/dto-validator/dto-validator.provider.ts
@@ -2,7 +2,7 @@ import "reflect-metadata";
 
 import { Request, Response, NextFunction, RequestHandler } from "express";
 import { StatusCode } from "../../error/status-code";
-import { packageResolver } from "../../common/package-resolver.provider";
+import { packageResolver } from "../../common/package-resolver";
 import { Logger } from "../logger/logger.provider";
 
 /**
